+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 - Refactor of Editor code
 - Removed "Buffers" from Scripts
+- Refactor of IsApproximately
 
 
 ## [Version: 1.9_FH] - 2019.04.22